requirements.txt
oarepo[s3]==12.0.21
cachetools==5.3.3
deepmerge==1.1.1
et-xmlfile==1.1.0
flask-babelex==0.9.4
frozendict==2.4.4
jinja2-simple-tags==0.6.1
jinjax==0.37
json5==0.9.25
langcodes==3.4.0
language-data==1.2.0
levenshtein==0.25.1
marisa-trie==1.2.0
marshmallow-i18n-messages==0.1.0
nr-metadata==2.0.32
nr-oaipmh-harvesters==1.0.20
nr-vocabularies==2.0.13
oarepo-communities==4.0.13
oarepo-dashboard==1.0.8
oarepo-global-search==1.0.15
oarepo-oai-pmh-harvester==4.0.46
oarepo-published-service==1.0.3
oarepo-requests==1.2.2
oarepo-runtime==1.5.43
oarepo-ui==5.1.41
oarepo-vocabularies==2.0.92
oarepo-workflows==1.0.1
openpyxl==3.1.5
python-dotenv==1.0.1
rapidfuzz==3.9.4
sickle==0.7.0
tqdm==4.66.4
whitenoise==6.7.0